==> notebooks/0_enst_to_gene_symbol_converter.ipynb <==
# ENST to gene symbol converter

import mygene
mg = mygene.MyGeneInfo()
import pandas as pd
import os,glob

enst_ids = pd.read_table('../data/raw/rna/HEK293T/SRR3997504.tsv')
enst_ids['enst'] = enst_ids.target_id.apply(lambda x: x.split('.')[0])

enst_ids['enst'].unique().shape,enst_ids['enst'].shape

%%time
enst_sym_df = pd.DataFrame(mg.querymany(enst_ids['enst'], scopes='ensembl.transcript', species='human'))

enst_sym_df.to_csv('../data/external/enst_sym_df.csv')

print(enst_sym_df.shape)
enst_sym_df[:5]

enst_sym_df.symbol.unique().shape
