==> notebooks/1A_mg_preprocess_rna.ipynb <==
preprocess rna

06/02/2020

save to `../data/interim/rna/

- `sample_tpm_sym.csv`
- `tissue_tpm_sym.csv`

import pandas as pd
import os,glob

save_dir = '../data/interim/rna/'
if not os.path.exists(save_dir):
    os.makedirs(save_dir)

tissue = ['Astrocytes','SL_D0','SL_D2', 'SLC_D0','SLC_D2','H9_D0', 'H9_D2', 'H9_D10', 'H9_D28']

rna_df = pd.read_excel('../data/raw/rna/H9SLSLC_0d_2d_28d_TPM.xlsx')
rna_d10 = pd.read_excel('../data/raw/rna/h9 d10 TPM.xlsx')
astro1_df = pd.read_csv('../data/raw/rna/epithelial_tpm_sym.csv',index_col=0)[['Astrocytes-B1','Astrocytes-B2']]

rna_d10_sel = rna_d10[['gene_id','TPM']]
rna_d10_sel.columns = ['gene_id', 'H9_D10']

rna_df_merged = rna_df.merge(rna_d10_sel, how='left',on='gene_id').merge(astro1_df,how='left',left_on='gene_id',right_index=True).fillna(0)
rna_df_merged.set_index('gene_id',inplace=True)

rna_df_merged.to_csv(os.path.join(save_dir,'sample_tpm_sym.csv'))

rna_df_merged.columns

rna_df_merged_tissue = rna_df_merged.copy()
rna_df_merged_tissue['Astrocytes'] = rna_df_merged_tissue[['Astrocytes-B1','Astrocytes-B2']].mean(axis=1)
rna_df_merged_tissue['SL_D0'] = rna_df_merged_tissue[['SL6d_1', 'SL6d_2', 'SL6d_3']].mean(axis=1)
rna_df_merged_tissue['SL_D2'] = rna_df_merged_tissue[['SL_Ngn2_2d_1', 'SL_Ngn2_2d_2', 'SL_Ngn2_2d_3']].mean(axis=1)
rna_df_merged_tissue['SLC_D0'] = rna_df_merged_tissue[['SLC6d_1', 'SLC6d_2','SLC6d_3']].mean(axis=1)
rna_df_merged_tissue['SLC_D2'] = rna_df_merged_tissue[['SLC_Ngn2_2d_1', 'SLC_Ngn2_2d_2', 'SLC_Ngn2_2d_3']].mean(axis=1)
rna_df_merged_tissue['H9_D0'] = rna_df_merged_tissue[['H9_1', 'H9_2']].mean(axis=1)
rna_df_merged_tissue['H9_D2'] = rna_df_merged_tissue[['H9_Ngn2_2d_1', 'H9_Ngn2_2d_2', 'H9_Ngn2_2d_3']].mean(axis=1)
rna_df_merged_tissue['H9_D10'] = rna_df_merged_tissue[['H9_D10']].mean(axis=1)
rna_df_merged_tissue['H9_D28'] = rna_df_merged_tissue[['H9_28d_BGI_1', 'H9_28d_BGI_2']].mean(axis=1)
rna_df_merged_tissue = rna_df_merged_tissue[tissue]

rna_df_merged_tissue.to_csv(os.path.join(save_dir, 'tissue_tpm_sym.csv'))

rna_df_merged_tissue = pd.read_csv(os.path.join(save_dir, 'tissue_tpm_sym.csv'),index_col=0)
rna_df_merged_tissue.sum(axis=0)/100000
