==> notebooks/2C_mg_hichip_group_diff_analysis.ipynb <==
edited 08/20/2020

import os, glob
import pandas as pd


normal_tissues = ['Astrocytes','SL_D0','SL_D2','SLC_D0', 'SLC_D2','H9_D0','H9_D2','H9_D10','H9_D28']

def get_alpha(row):
    return '::'.join(sorted([row['source'], row['target']]))

%%time
# takes a couple min
tissue_loop_dict = {}
for file in sorted(glob.glob('../data/interim/merged/loops/*csv')):
    filename = os.path.basename(file)
    tissue = filename.split('.')[0]
    if tissue in normal_tissues:
        print(file)
        df = pd.read_csv(file, index_col=0)
        df['loop_name'] = df[['source','target']].apply(get_alpha,axis=1)
        per_tissue_loop_dict = pd.Series(df['count'].values, index=df.loop_name.values).to_dict()
        tissue_loop_dict[tissue] = per_tissue_loop_dict
        

tissue_loop_df = pd.DataFrame(tissue_loop_dict).fillna(0)

save_dir = '../data/processed/fig1/hichip'
if not os.path.exists(save_dir):
    os.makedirs(save_dir)

tissue_loop_df.to_csv(os.path.join(save_dir,'tissue_loop_df.csv'))
