==> notebooks/2E_pairwise_comp.ipynb <==
# 2E_pairwise_comp

import os,glob
import pandas as pd
import pickle
import numpy as np
import scipy.stats as stats

rna_genes_files = glob.glob('../data/processed/fig1/rna/rna_genes*.csv')
print(rna_genes_files)
hichip_genes_files = glob.glob('../data/processed/fig1/hichip/time_time_genes_*.csv')
print(hichip_genes_files)
atac_gene_files =['../data/processed/fig1/atac/atac_time_genes_H9_D0.csv', '../data/processed/fig1/atac/atac_time_genes_H9_D2.csv',
                 '../data/processed/fig1/atac/atac_time_genes_H9_D10.csv','../data/processed/fig1/atac/atac_time_genes_H9_D28.csv']
#glob.glob('../data/processed/fig1/atac/atac_time_genes*.csv')
atac_gene_files

save_dir = '../data/processed/fig1/'

def get_genes(gene_files):
    all_genes_bg = []
    all_gene_dict = {}
    for file in gene_files:
        filename = os.path.basename(file).split('.')[0]
        tissue = '_'.join(filename.split('_')[-2:] )## modify
        genes = list(pd.read_csv(file,header=None).loc[:,0])
        all_gene_dict[tissue] = sorted(set(genes))
        all_genes_bg+=all_gene_dict[tissue] 
    
    all_gene_dict['background'] = sorted(set(all_genes_bg))
    return all_gene_dict
        
rna_genes = get_genes(rna_genes_files)
hichip_genes = get_genes(hichip_genes_files)
atac_genes = get_genes(atac_gene_files)

atac_genes.keys()

def comp_two_gene_dicts(gene_dict1, gene_dict2, comp_groups, pseudo=0):
    """
    1: rows
    2: columns
    """
    bg = set(gene_dict1['background']) | set(gene_dict2['background'])
    
    results = {}
    result_idx = 0
    for color_1 in comp_groups:
        for color_2 in comp_groups:
            geneset_1 = set(gene_dict1[color_1])
            geneset_2 = set(gene_dict2[color_2])
            
            
            in_1_and_in_2 = geneset_1 & geneset_2
            in_1_not_2 = geneset_1 - geneset_2
            not_1_in_2 = geneset_2 - geneset_1
            in_1_or_in_2 = geneset_1 | geneset_2
            not_1_not_2 = bg - in_1_or_in_2
            
            mat_counts = np.array([[len(in_1_and_in_2), len(in_1_not_2)],
                                  [len(not_1_in_2), len(not_1_not_2)]]).reshape((2,2))
            mat_counts_pseudo = mat_counts+pseudo
#             if (color_1=='grey') & (color_2=='grey'):
#                 print(mat_counts)
    
            num_in_1 = mat_counts.sum(axis=1)[0]
            num_in_2 = mat_counts.sum(axis=0)[0]
            observed_num = mat_counts[0][0]
            expected_num = num_in_1*num_in_2/sum(sum(mat_counts))
            oddsratio_pseudo, pvalue_pseudo = stats.fisher_exact(mat_counts_pseudo,alternative='greater')
            jaccard = len(in_1_and_in_2)/len(in_1_or_in_2)
            intersect_over_min = len(in_1_and_in_2)/min(num_in_1,num_in_2)
            
            results[result_idx] = {'color_row':color_1, 'color_col':color_2,
                                   'jaccard':jaccard,'intersect_over_min':intersect_over_min,
                                    'intersection':len(in_1_and_in_2), 
                                   'union':len(in_1_or_in_2), 
                                   'num_in_1':num_in_1,'num_in_2':num_in_2,
                                   'observed':observed_num, 'expected':expected_num, 
                                   'oddsratio':oddsratio_pseudo, 'pval':pvalue_pseudo}
    
            result_idx+=1
    result_df = pd.DataFrame.from_dict(results,orient='index')
    result_df['pval_bonf'] = result_df.pval.apply(lambda x: min(1, x* sum(sum(mat_counts))))#result_df.shape[0]))
    result_df['log_pval_bonf'] = result_df.pval_bonf.apply(lambda x: min(100,-np.log10(x+1e-100)))
    return result_df

comp_val = ['H9_D0', 'H9_D2', 'H9_D10', 'H9_D28']

rna_hichip_df = comp_two_gene_dicts(rna_genes, hichip_genes,comp_groups=comp_val)
rna_hichip_df.to_csv(os.path.join(save_dir, 'rna_hichip_df.csv'))

rna_atac_df = comp_two_gene_dicts(rna_genes, atac_genes,comp_groups=comp_val)
rna_atac_df.to_csv(os.path.join(save_dir, 'rna_atac_df.csv'))

atac_hichip_df = comp_two_gene_dicts(atac_genes, hichip_genes,comp_groups=comp_val,pseudo=1)
atac_hichip_df.to_csv(os.path.join(save_dir, 'atac_hichip_df.csv'))

rna_hichip_df

stats.fisher_exact([[ 2067,  5227],
 [  620, 10668]])

rna_atac_df.pivot(index='color_row',columns='color_col',values='log_pval_bonf')

rna_atac_df.pivot(index='color_row',columns='color_col',values='jaccard')

rna_atac_df.pivot(index='color_row',columns='color_col',values='intersect_over_min')
