==> notebooks/9A_make_mpra_tracks.ipynb <==
# 9A_make_mpra_tracks


2/16/2021

how to make rna tracks : will be displaying TPM values on exons in hg19

Option 1: Wiggle (not done b/c wiggle not easy to process)
1. make wiggle file (https://genome.ucsc.edu/goldenPath/help/wiggle.html)
2. convert wig to BigWig (https://genome.ucsc.edu/goldenpath/help/bigWig.html)

Option 2: Bedgraph
1. make bedgraph (https://genome.ucsc.edu/goldenpath/help/bedgraph.html)
2. convert bedgraph to BigWig (https://genome.ucsc.edu/goldenpath/help/bigWig.html)
    To create a bigWig track from a bedGraph file, follow these steps:

1. Create a bedGraph format file following the directions here. When converting a bedGraph file to a bigWig file, you are limited to one track of data in your input file; therefore, you must create a separate bedGraph file for each data track.
2. Remove any existing track or browser lines from your bedGraph file so that it contains only data.
3. Download the bedGraphToBigWig program from the binary utilities directory.
4. Use the fetchChromSizes script from the same directory to create the chrom.sizes file for the UCSC database with which you are working (e.g., hg19). If the assembly genNom is hosted by UCSC, chrom.sizes can be a URL like http://hgdownload.soe.ucsc.edu/goldenPath/genNom/bigZips/genNom.chrom.sizes
5. Use the bedGraphToBigWig utility to create a bigWig file from your bedGraph file:
    `bedGraphToBigWig in.bedGraph chrom.sizes myBigWig.bw`
(Note that the bedGraphToBigWig program DOES NOT accept gzipped bedGraph input files.)
6. Move the newly created bigWig file (myBigWig.bw) to a web-accessible http, https, or ftp location.
7. Paste the URL into the custom track entry form or construct a custom track using a single track line.
8. Paste the custom track line into the text box on the custom track management page.



import pandas as pd
import os, glob, re
import subprocess
import pybedtools
import numpy as np

# 0. get files
- exon file
- rna seq tpm matrix file
- chrom size file

save_dir = '../data/processed/mpra_bigwigs'
if not os.path.exists(save_dir):
    os.makedirs(save_dir)
if not os.path.exists(os.path.join(save_dir, 'bedgraphs')):
    os.makedirs(os.path.join(save_dir, 'bedgraphs'))
if not os.path.exists(os.path.join(save_dir, 'bigwigs')):
    os.makedirs(os.path.join(save_dir, 'bigwigs'))

size_file = '../data/external/chrom_hg19.sizes'
exon_bed = '../data/external/gencode.v19.exon.bed'

exon_df = pybedtools.BedTool(exon_bed).to_dataframe()
exon_df[:5]

# rna_file = '../data/interim/rna/tissue_tpm_sym.csv'
# rna_df = pd.read_csv(rna_file)
# print(rna_df.shape)
# rna_df[:5]

mpra_df = pd.read_excel('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/figure_tables/lib_df_annon_mprasig_FINAL_v020521.xlsx')
mpra_df_sel = mpra_df[['rowname','Chr','Position','Ref/Alt','Linked_SNP','diseases',
        'mpra_tissue',	'mpra_pval_str',	'mpra_pval_mean',
         'mpra_logfc_str',	'mpra_logfc_mean',	'bool_mpra_sig']]
mpra_df_sel['end'] = mpra_df_sel.Position +1
mpra_df_sel['neglog10pval'] = mpra_df_sel.mpra_pval_mean.apply(lambda x:-np.log10(x))
mpra_df_sel = mpra_df_sel[['Chr','Position','end','neglog10pval', 'mpra_logfc_mean','mpra_tissue','Linked_SNP','diseases']]
mpra_df_sel.columns = ['chr','start','stop','name', 'mpra_logfc_mean','mpra_tissue','Linked_SNP','diseases']
mpra_df_sel.to_csv(os.path.join(save_dir, 'mpra_df_sel.csv'))
mpra_df_sel[:5]

tissues = ['AST', 'SL',  'SLC','H9D0', 'H9D2', 'H9D4','H9D10','HEK293']

# # bedgraph defintion lines
# definition_str = "track type=bedGraph name={track_label:s} description=center_label visibility=display_mode color=r,g,b altColor=r,g,b priority=priority autoScale=on|off alwaysZero=on|off gridDefault=on|off maxHeightPixels=max:default:min graphType=bar|points viewLimits=lower:upper yLineMark=real-value yLineOnOff=on|off windowingFunction=maximum|mean|minimum smoothingWindow=off|2-16"
# definition_str.format(track_label='test', )

# txt = "For only {price:.2f} dollars!"
# print(txt.format(price = 49))

# 1A. make bedGraph (no brower or track lines) per tissue

space delimited
`chrom chromStart chromEnd dataValue`

example:

```
chr19 49302000 49302300 -1.0
chr19 49302300 49302600 -0.75
```

for tissue in tissues:
    print('*****', tissue)
    mpra_df_tissue = mpra_df_sel[mpra_df_sel.mpra_tissue.str.contains(tissue)]
    
    mpra_df_tissue = mpra_df_tissue[['chr','start','stop','name']]
    mpra_df_bed = pybedtools.BedTool.from_dataframe(mpra_df_tissue).sort()#.merge(c=4,o='sum')
    mpra_df_tissue = mpra_df_bed.to_dataframe()
    print('size mpra_df_bed sorted ',mpra_df_tissue.shape)
    mpra_df_tissue.to_csv(os.path.join(save_dir,'bedgraphs', tissue+'_unsorted.bedGraph'),index=False, header=False,sep=' ')    
    

mpra_df_tissue[:5]

# 1B. sort bedGraphs

`sort -k1,1 -k2,2n unsorted.bedGraph > sorted.bedGraph`

for tissue in tissues:
    unsort_file = os.path.join(save_dir, 'bedgraphs',tissue+'_unsorted.bedGraph')
    sort_file = os.path.join(save_dir,'bedgraphs', tissue+'_sorted.bedGraph')
    sort_cmd = 'sort -k1,1 -k2,2n "{}" > "{}"'.format(unsort_file,sort_file)
    print(sort_cmd)
    subprocess.call(sort_cmd,shell=True)

# 2. run `bedGraphToBigWig`
cmd: `bedGraphToBigWig in.bedGraph chrom.sizes myBigWig.bw`

bedGraphToBigWig_cmd = '/Users/mguo123/Documents/genome_browser_tools/bedGraphToBigWig'


for tissue in tissues:
    bedgraph_file = os.path.join(save_dir,'bedgraphs', tissue+'_sorted.bedGraph')
    bigwig_file = os.path.join(save_dir,'bigwigs', tissue+'.bw')
    cmd = '{} {} {} {}'.format(bedGraphToBigWig_cmd,bedgraph_file, size_file,bigwig_file )
    print(cmd)
    subprocess.call(cmd,shell=True)

# 3. make hub.config.json

manually made

final folder for upload: `/Users/mguo123/Documents/pan_omics_psych/data/processed/mpra_bigwigs/bigwigs`

# 4. checking
