==> notebooks/10A_compare_atac.ipynb <==
# 10A_compare atac to other atac papers

- natgen song data
- duan paper science ASoC
- inoue 2019 cell stem cell

import pandas as pd
import pybedtools
import os, glob
from collections import defaultdict

save_dir = '../data/processed/compare_external/song_atac/'

def get_stats(feature_dict1, feature_dict2, save_prefix, 
              save_bed = False, # save bed A
              save_dir=save_dir):
    stats_dict = {}
    stats_idx = 0
    for my_sampleA, my_bedA in feature_dict1.items():
        num_peaksA = len(my_bedA)
        for my_sampleB, my_bedB in feature_dict2.items():
            num_peaksB = len(my_bedB)
            if save_bed:
                new_name = '__'.join([my_sampleA, my_sampleB,'intersect'])
                my_bedA.intersect(my_bedB,wa=True,f=0.25).saveas(os.path.join(save_dir, new_name+'.bed'))
            stats_dict[stats_idx] = {'sampleA':my_sampleA,
                                    'sampleB':my_sampleB,
                                    'num_peaksA':num_peaksA,
                                    'num_peaksB':num_peaksB}
            stats_dict[stats_idx].update(my_bedA.jaccard(b=my_bedB,f=0.25))
            stats_idx+=1

    stats_df = pd.DataFrame.from_dict(stats_dict,orient='index')
    stats_df['perc_peaksA_captured'] = stats_df.n_intersections/stats_df.num_peaksA
    stats_df['perc_peaksB_captured'] = stats_df.n_intersections/stats_df.num_peaksB
    stats_df.to_csv(os.path.join(save_dir, 'stats_'+save_prefix+'_comp.csv'))   
    return stats_df



my_atac_dir = '../data/interim/merged/atac/'
my_atac_files = glob.glob(os.path.join(my_atac_dir, '*'))
my_atac_files_named = {}
for file in my_atac_files:
    sample = os.path.basename(file).split('_merged')[0]
    df = pybedtools.BedTool(file).sort().merge().to_dataframe()
    df['name']= ['Peak_'+str(i+1) for i in range(len(df))]
    my_atac_files_named[sample] = pybedtools.BedTool.from_dataframe(df)
my_atac_files, my_atac_files_named.keys() 

# pybedtools.BedTool(my_atac_files[0]).to_dataframe()[:5]

song_files =glob.glob('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/song_atac/*.narrowPeak')
song_files_named = {}
for file in song_files:
    sample = os.path.basename(file).split('_')[1].split('.')[0]
    df = pybedtools.BedTool(file).sort().merge().to_dataframe()
    df['name']= ['Peak_'+str(i+1) for i in range(len(df))]
    song_files_named[sample] = pybedtools.BedTool.from_dataframe(df)
    
    
song_files,song_files_named.keys()

%%time
stats_df_song_my = get_stats(my_atac_files_named, song_files_named, 
                             save_bed=True, save_dir = '../data/processed/compare_external/song_atac/',
                             save_prefix='my_song')
stats_df_song_my.sort_values('perc_peaksA_captured',ascending=True)


# %%time
# stats_dict = {}
# stats_idx = 0
# for my_sample, my_bed in my_atac_files_named.items():
#     num_my_peaks = len(my_bed)
#     for song_sample, song_bed in song_files_named.items():
#         num_song_peaks = len(song_bed)
#         new_name = '__'.join([my_sample, song_sample,'intersect'])
# #         print(new_name)
#         my_bed.intersect(song_bed,wa=True).saveas(os.path.join(save_dir, new_name+'.bed'))
#         stats_dict[stats_idx] = {'my_sample':my_sample,
#                                 'song_sample':song_sample,
#                                 'num_my_peaks':num_my_peaks,
#                                 'num_song_peaks':num_song_peaks}
#         stats_dict[stats_idx].update(my_bed.jaccard(b=song_bed))
#         stats_idx+=1
        

%%time
stats_df_my_my = get_stats(my_atac_files_named, my_atac_files_named, 
                             save_bed=False, save_dir = '../data/processed/compare_external/song_atac/',
                             save_prefix='my_my')
stats_df_my_my.sort_values('perc_peaksA_captured',ascending=True)


%%time
stats_df_song_song = get_stats(song_files_named, song_files_named, 
                             save_bed=False, save_dir = '../data/processed/compare_external/song_atac/',
                             save_prefix='song_song')
stats_df_song_song.sort_values('perc_peaksA_captured',ascending=True)


takeaways - astro/"cytes seem similar to astrocyts 

the neuron cells are most similar to atac H9D10 and D28
# to do make plot

# 2. duan stuff

convert hg38 to 19


# sort merge then then convert to hg19
duan_files =glob.glob('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/aay3983_datas1_duan_allele_specific/*.txt')
duan_files_named = {}
for file in duan_files:
    sample = os.path.basename(file).split('.')[0]
    print(sample)
    df = pd.read_table(file)
    df.columns = ['chr','start','end','name','score','strand','summit']
    print(df.shape, 'before')
    df = pybedtools.BedTool.from_dataframe(df).sort().merge().to_dataframe()
    print(df.shape, 'after')
    df['name']= ['Peak_'+str(i+1) for i in range(len(df))]
    bed = pybedtools.BedTool.from_dataframe(df)
    bed.saveas('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/aay3983_datas1_duan_allele_specific/'+sample+'_hg38.bed')
    duan_files_named[sample] = bed
    
    
duan_files,duan_files_named.keys()

# sort merge then then convert to hg19
duan_files =glob.glob('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/aay3983_datas1_duan_allele_specific/*_hg19.bed')
duan_files_named = {}
for file in duan_files:
    sample = os.path.basename(file).split('.')[0].split('_hg')[0]
    print(sample)
    df = pybedtools.BedTool(file).sort().merge().to_dataframe()
    print(df.shape, 'hg19')
    df['name']= ['Peak_'+str(i+1) for i in range(len(df))]
    duan_files_named[sample] = pybedtools.BedTool.from_dataframe(df)
    
    
duan_files,duan_files_named.keys()

%%time
stats_df_duan = get_stats(duan_files_named, duan_files_named, save_prefix='duan_duan')
stats_df_duan.sort_values('perc_peaksA_captured',ascending=True)


%%time
stats_df_duan_my = get_stats(my_atac_files_named, duan_files_named, 
                             save_bed=True, save_dir = '../data/processed/compare_external//duan_atac/',
                             save_prefix='my_duan')
stats_df_duan_my.sort_values('perc_peaksA_captured',ascending=True)


# 3. inoue 2019 cell stem cell

['/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/inoue_atac/GSM3163878_ATACseq_7TP_24hr_rep1_peaks.narrowPeak.gz',


inoue_files = glob.glob('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/inoue_atac/*')
inoue_rep_bed_dict = defaultdict(list)
for file in inoue_files:
    _, _, _, time, rep, _ = os.path.basename(file).split('_')
    print(time, rep)
    inoue_rep_bed_dict[time].append(pybedtools.BedTool(file).sort().merge())
    
    
    
    

inoue_files_named = {}
for time, bed_arr in inoue_rep_bed_dict.items():
    print(time)
    bed_merge = bed_arr[0] + bed_arr[1]
    df = bed_merge.to_dataframe()
    print(df.shape, 'hg19')
    df['name']= ['Peak_'+str(i+1) for i in range(len(df))]
    inoue_files_named[time] = pybedtools.BedTool.from_dataframe(df)
    

inoue_files_named.keys()

%%time
stats_df_inoue = get_stats(inoue_files_named, 
                          inoue_files_named, save_prefix='inoue_inoue')
stats_df_inoue.sort_values('perc_peaksA_captured',ascending=True)


%%time
stats_df_inoue_my = get_stats(my_atac_files_named, inoue_files_named, 
                             save_bed=True, save_dir = '../data/processed/compare_external/inoue_atac/',
                             save_prefix='my_inoue')
stats_df_inoue_my.sort_values('perc_peaksA_captured',ascending=True)
