==> notebooks/10B_compar_hichip.ipynb <==
# compare hichip

song
inoue?


import pandas as pd
import pybedtools
import os, glob
from collections import defaultdict

save_dir = '../data/processed/compare_external/song_hichip/'
if not os.path.exists(save_dir):
    os.makedirs(save_dir)

def get_stats(feature_dict1, feature_dict2, save_prefix, 
              save_bed = False, # save bed A
              save_dir=save_dir):
    stats_dict = {}
    stats_idx = 0
    for my_sampleA, my_bedA in feature_dict1.items():
        num_anchorsA = len(my_bedA)
        for my_sampleB, my_bedB in feature_dict2.items():
            num_anchorsB = len(my_bedB)
            if save_bed:
                new_name = '__'.join([my_sampleA, my_sampleB,'intersect'])
                my_bedA.intersect(my_bedB,wa=True,f=0.25).saveas(os.path.join(save_dir, new_name+'.bed'))
            stats_dict[stats_idx] = {'sampleA':my_sampleA,
                                    'sampleB':my_sampleB,
                                    'num_anchorsA':num_anchorsA,
                                    'num_anchorsB':num_anchorsB}
            stats_dict[stats_idx].update(my_bedA.jaccard(b=my_bedB,f=0.25))
            stats_idx+=1

    stats_df = pd.DataFrame.from_dict(stats_dict,orient='index')
    stats_df['perc_anchorsA_captured'] = stats_df.n_intersections/stats_df.num_anchorsA
    stats_df['perc_anchorsB_captured'] = stats_df.n_intersections/stats_df.num_anchorsB
    stats_df.to_csv(os.path.join(save_dir, 'stats_'+save_prefix+'_comp.csv'))   
    return stats_df



my_anchor_dir = '../data/interim/merged/anchors_bed_sort//'
my_anchor_files = glob.glob(os.path.join(my_anchor_dir, '*'))
my_anchor_files_named = {}
for file in my_anchor_files:
    sample = os.path.basename(file).split('_sort')[0]
    my_anchor_files_named[sample] = pybedtools.BedTool(file)
    print(sample, len(my_anchor_files_named[sample]))
my_anchor_files, my_anchor_files_named.keys() 

my_loop_files = glob.glob('../data/interim/merged/loops/*')
my_loop_files_named = {}
for file in my_loop_files:
    sample = os.path.basename(file).split('.')[0]
    my_loop_files_named[sample] = pd.read_csv(file,index_col=0)
    print(sample, my_loop_files_named[sample].shape)


my_loop_files_named['SL_D0']

song_pcdir = '/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/novogene_071420/external_data/song_pchic'
song_loop_files = glob.glob(os.path.join(song_pcdir, '*txt'))

song_anchor_files_named = {}

song_loop_files_named = {}
for file in song_loop_files:
    sample = os.path.basename(file).split('.')[0].split('_')[1]

    df = pd.read_table(file,names=['source','target','count'])
    df[['chrx','startx','stopx']]  = df.source.str.split(',',expand=True)
    df[['chry','starty','stopy']]  = df.target.str.split(',',expand=True)
    df['startx_round'] = (round(df['startx'].map(int) / 5000) * 5000).map(int)
    df['stopx_round_prelim'] = (round(df['stopx'].map(int) / 5000) * 5000)
    df['stopx_round'] = df['startx_round']+5000
    df['stopx_round'] = df[['stopx_round_prelim','stopx_round']].max(axis=1).map(int)
    df['starty_round'] = (round(df['starty'].map(int) / 5000) * 5000).map(int)
    df['stopy_round_prelim'] = (round(df['stopy'].map(int) / 5000) * 5000)
    df['stopy_round'] = df['starty_round']+5000
    df['stopy_round'] = df[['stopy_round_prelim','stopy_round']].max(axis=1).map(int)
    df['source_arr'] = df.apply(lambda row: [row.chrx+'_'+str(start) + '_'+ str(start+5000) for start in list(range(row.startx_round, row.stopx_round, 5000))],axis=1)
    df['target_arr'] = df.apply(lambda row: [row.chry+'_'+str(start) + '_'+ str(start+5000) for start in list(range(row.starty_round, row.stopy_round, 5000))],axis=1)

    df = df[['source_arr','target_arr','count']].explode('source_arr').explode('target_arr')
    df.columns = ['source','target', 'count']
    
    df.to_csv(os.path.join(song_pcdir, sample+'.loops.csv'))
                            
    anc_df = pd.DataFrame(pd.concat([df.source , df.target]).unique(),columns=['name'])
    anc_df[['chr','start','stop']] = anc_df.name.str.split('_',expand=True)
    anc_df = anc_df[['chr', 'start', 'stop', 'name'   ]]
    anc_bed = pybedtools.BedTool.from_dataframe(anc_df).sort()
    anc_bed.saveas(os.path.join(song_pcdir, sample+'_sort.bed'))
                            
                            
    song_loop_files_named[sample] = df
    song_anchor_files_named[sample] = anc_bed
                            
    print(sample, song_loop_files_named[sample].shape, len(anc_bed))

%%time
# compare anchors - similar to atac peaks
stats_df_song_my = get_stats(my_anchor_files_named, song_anchor_files_named, 
                             save_bed=True, 
                             save_prefix='my_song')
    

stats_df_song_my.sort_values('perc_anchorsB_captured',ascending=True)


# compare loops just report # loops that overlap
stats_dict = {}
stats_idx = 0
for sampleA, loop_dfA in my_loop_files_named.items():
    loop_dfA_arr = set(loop_dfA.source +'::' + loop_dfA.target)
    for sampleB, loop_dfB in song_loop_files_named.items():
        loop_dfB = loop_dfB[['source','target']].drop_duplicates()
        loop_dfB_arr = set(loop_dfB.source +'::' + loop_dfB.target)
        num_loopA = len(loop_dfA_arr)
        num_loopB = len(loop_dfB_arr)
        num_intersection = len(loop_dfA_arr.intersection(loop_dfB_arr))
        num_union = len(loop_dfA_arr.union(loop_dfB_arr))
        
        stats_dict[stats_idx] = {'sampleA':sampleA,
                                'sampleB':sampleB,
                                'num_loopA':num_loopA,
                                'num_loopB':num_loopB,
                                'num_intersection_loops':num_intersection,
                                'num_union':num_union}
        stats_idx+=1

stats_df_song_loop = pd.DataFrame.from_dict(stats_dict,orient='index')
stats_df_song_loop['perc_loopsA_captured'] = stats_df_song_loop.num_intersection_loops/stats_df_song_loop.num_loopA
stats_df_song_loop['perc_loopsB_captured'] = stats_df_song_loop.num_intersection_loops/stats_df_song_loop.num_loopB

stats_df_song_loop.to_csv(os.path.join(save_dir, 'stats_'+'my_song_loop'+'_comp.csv'))   


stats_df_song_loop.sort_values('')

loop_dfA.merge(loop_dfB, how='outer',on=['source','target']).shape
