==> notebooks/12_crm_target_mpra_genes.ipynb <==
# CRM target mpra genes

Point 14: “It would be informative to include a new figure including a gene ontology and/or pathway analysis of all TFs identified from the motif analysis. This would represent an informative big picture analysis of all pathways associated with psychiatric disease. Analysis of TF motifs disrupted by the identified daSNVs should include RNA-seq (or qPCR), demonstrating the relative expression of these TFs in neurons and astrocytes. Are top TFs expressed in the fetal/adult post-mortem brain? Perhaps a cell-type-specific or developmental analysis of TF and target gene expression might inform when these different aspects of disease risk exert their effects? Does transcriptomic imputation (e.g.TWAS/PrediXcan) predict an enrichment of TF gene expression in psychiatric disease?”

With regards to the linking TF target genes, we will link these TFs to putative target genes by via our matched RNA, ATAC, and HiChIP dataset, and finally complete the analysis by describing neuropsychiatric disease relevance of these TF targeted genes. These analyses greatly strengthen the main messages of the work. 

import pandas as pd
import numpy as np
import os, glob, sys

tfs = pd.read_csv('../data/external/HOCOMOCOv11_annotation.csv').tf.values
tfs

all_mpra_egenes_mpra = pd.read_table('/Users/mguo123/Google Drive/1_khavari/noncancer_project/miseq/MPRA_022022/mpraanalyze/egene_gtex/all.txt', header=None)
all_mpra_egenes_mpra.columns = ['genes']
all_mpra_egenes_mpra

all_mpra_egenes_mpra['is_tf'] = all_mpra_egenes_mpra.genes.apply(lambda x: x in tfs)
all_mpra_egenes_mpra

mpra_tf_list = all_mpra_egenes_mpra.genes.to_list()
mpra_tf_list_col = []
for tf in mpra_tf_list:
    mpra_tf_list_col.append(tf+'_pro')
    mpra_tf_list_col.append(tf+'_loop')

all_mpra_egenes_mpra.is_tf.value_counts()

mpra_df = all_mpra_egenes_mpra[all_mpra_egenes_mpra.is_tf]

%%time
crm_df = pd.read_csv("../data/processed/tissue_crms/all_count_comb_overall.csv")
crm_df.set_index('Unnamed: 0',inplace=True)

mpra_tf_list_col_filt = [x for x in mpra_tf_list_col if x in crm_df.columns]
mpra_tf_list_filt = sorted(set([x.split('_')[0] for x in mpra_tf_list_col_filt]))
print(mpra_tf_list_filt)
mpra_tf_list_col_filt

crm_df[:5]

print(crm_df.shape)
crm_df.loc[:,mpra_tf_list_col_filt]

mpra_tf_list_filt

for tf in mpra_tf_list_filt:
    print('((()))')
    print(tf)
    genes_target_pro = list(set(crm_df.index[(crm_df[tf+'_pro']>0) & (crm_df.exp>1)]))
    genes_target_loop = list(set(crm_df.index[(crm_df[tf+'_loop']>0) & (crm_df.exp>1)]))
    print(len(genes_target_pro),len(genes_target_loop))
    with open('../data/processed/mpra_egenes/'+tf+'.txt','w') as f:
        for x in genes_target_pro:
            f.write(x+'\n')
#     for x in genes_target_pro:
#         print(x)
#     raise
